--- src/templates/index.html
@@ -1,60 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>DEEP FROST</title>
-
-    <!-- CSS only -->
-    <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.5.2/css/bootstrap.min.css"
-        integrity="sha384-JcKb8q3iqJ61gNV9KGb8thSsNjpSL0n8PARn9HuZOnIxN0hoP+VmmDGMN5t9UJ0Z" crossorigin="anonymous">
-
-    <!-- JS, Popper.js, and jQuery -->
-    <script src="https://code.jquery.com/jquery-3.5.1.slim.min.js"
-        integrity="sha384-DfXdz2htPH0lsSSs5nCTpuj/zy4C+OGpamoFVy38MVBnE+IbbVYUew+OrCXaRkfj"
-        crossorigin="anonymous"></script>
-    <script src="https://cdn.jsdelivr.net/npm/popper.js@1.16.1/dist/umd/popper.min.js"
-        integrity="sha384-9/reFTGAW83EW2RDu2S0VKaIzap3H66lZH81PoYlFhbGU+6BZp6G7niu735Sk7lN"
-        crossorigin="anonymous"></script>
-    <script src="https://stackpath.bootstrapcdn.com/bootstrap/4.5.2/js/bootstrap.min.js"
-        integrity="sha384-B4gt1jrGC7Jh4AgTPSdUtOBvfO8shuf57BaghqFfPlYxofvL8/KUEfYiJOMMV+rV"
-        crossorigin="anonymous"></script>
-</head>
-
-<body>
-    <div class="container">
-        <div class="row">
-        <h1 class="mx-auto pt-2">DEEP FROST</h1>
-        </div>
-        <hr>
-        <div class="row pt-2">
-        <div class="col-4 border">
-            <form action="/" method="POST" class="p-3">{{ form.hidden_tag() }}
-                {% for field in form if field.widget.input_type != 'hidden' %}
-                <div class="form-group">
-                    <strong>{{ field.label(class='control-label') }}</strong>
-                    {{ field(class='form-control') }}
-                    {% if field.errors %}
-                    <div class="alert alert-danger" role="alert">
-                        {{ field.errors[0] }}
-                    </div>
-                    {% endif %}
-                </div>
-                {% endfor %}
-
-                <button type="submit" class="btn btn-primary ml-auto">Generate</button>
-            </form>
-        </div>
-        <div class="col-7 ml-auto">
-            {% if poem %}
-            {% for line in poem %}
-            <p>{{line|capitalize}}</p>
-            {% endfor %}
-            {% endif %}
-        </div>
-    </div>  
-    </div>
-</body>
-
-</html>
